# Adds benches for Serpent + questions regarding optimization