--- a/serpent/benches/serpent.rs
+++ b/serpent/benches/serpent.rs
@@ -1,0 +1,9 @@
+#![no_std]
+#![feature(test)]
+#[macro_use]
+extern crate block_cipher_trait;
+extern crate serpent;
+
+use serpent::Serpent;
+
+bench!(Serpent, 16);