--- a/serpent/src/lib.rs
+++ b/serpent/src/lib.rs
@@ -28,9 +28,7 @@ pub struct Serpent {
     k: Subkeys,
 }
 
-fn get_bit(x: usize, i: usize) -> u8 {
-    (x >> i) as u8 & 0x01
-}
+fn get_bit(x: usize, i: usize) -> u8 { (x >> i) as u8 & 0x01 }
 
 fn linear_transform_bitslice(input: Block128, output: &mut Block128) {
     let mut words = [0u32; 4];
@@ -71,7 +69,8 @@ fn round_bitslice(
     b_i: Block128,
     k: Subkeys,
     b_output: &mut Block128,
-) {
+)
+{
     let xored_block = xor_block(b_i, k[i]);
 
     let s_i = apply_s_bitslice(i, xored_block);
@@ -87,7 +86,8 @@ fn round_inverse_bitslice(
     b_i_next: Block128,
     k: Subkeys,
     b_output: &mut Block128,
-) {
+)
+{
     let mut s_i = [0u8; 16];
     if i == ROUNDS - 1 {
         s_i = xor_block(b_i_next, k[ROUNDS]);
@@ -100,9 +100,7 @@ fn round_inverse_bitslice(
     *b_output = xor_block(xored, k[i]);
 }
 
-fn apply_s(index: usize, nibble: u8) -> u8 {
-    S[index % 8][nibble as usize]
-}
+fn apply_s(index: usize, nibble: u8) -> u8 { S[index % 8][nibble as usize] }
 fn apply_s_inverse(index: usize, nibble: u8) -> u8 {
     S_INVERSE[index % 8][nibble as usize]
 }
@@ -229,6 +227,7 @@ impl BlockCipher for Serpent {
         Self::new_varkey(key).unwrap()
     }
 
+    #[inline(always)]
     fn new_varkey(key: &[u8]) -> Result<Self, InvalidKeyLength> {
         if key.len() < 16 || key.len() > 32 {
             return Err(InvalidKeyLength);
@@ -240,6 +239,7 @@ impl BlockCipher for Serpent {
         })
     }
 
+    #[inline(always)]
     fn encrypt_block(&self, block: &mut GenericArray<u8, Self::BlockSize>) {
         let mut b = [0u8; 16];
 
@@ -253,6 +253,7 @@ impl BlockCipher for Serpent {
         *block = *GenericArray::from_slice(&b);
     }
 
+    #[inline(always)]
     fn decrypt_block(&self, block: &mut GenericArray<u8, Self::BlockSize>) {
         let mut b = [0u8; 16];