I added benches for Serpent but results are not so satisfying, it could be x3-x10 better.

Here are the benchmark results :

```
running 2 tests
test decrypt ... bench:       8,447 ns/iter (+/- 200) = 1 MB/s
test encrypt ... bench:       8,325 ns/iter (+/- 676) = 1 MB/s
```

I have one simple question : could you guys provide some hints to improve the current Serpent implementation encrypt and decrypt speed? I don't mean Serpent specifics but more like Rust performance optimization such as `inline` and good practices.

Another question for you @tarcieri, earlier you told me about checking `branch instruction` on ASM, since I'm kinda newbie on this could you tell me what I have to consider regarding this?

Thanks!