@tarcieri 2020-03-13
@Slals I don't know much about Serpent, however those numbers certainly do seem quite slow