@jack-signal 2020-10-05
Right now it seems like Serpent is implemented in this crate using table lookups instead of the usual bitsliced sbox expressions. Changing that would probably provide about a 50x-100x performance improvement. In bitsliced mode it is also easy to use SIMD for parallel execution.