@tarcieri 2020-11-24
@Slals would you mind changing this PR to just include the benchmarks for now? Thanks!