@Slals 2020-12-02
Thanks @jack-signal for your feedback. I did't take the time to make the change in order to speed it. 

@tarcieri I removed the "inlines" tags if it's what you needed.