@tarcieri 2020-12-02
@Slals there are some conflicting changes in `serpent/src/lib.rs`. They might just go away if you run `rustfmt` on this branch. Otherwise you need to rebase.