@tarcieri 2021-01-11
I think the original goal of this PR was actually accomplished in #105 last May. Going to close this out.